requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-telegram-bot==20.7
aiohttp==3.9.1
apscheduler==3.10.4
python-dotenv==1.0.0
lxml==4.9.3
pytz==2023.3
brotli==1.1.0
feedparser==6.0.10 